完成子函数的编写，但是仍然有很多的需要更改的子函数。主要的难点在cc_decode改写，dct变换，MIMO_detectRB.c，最后是测试函数撰写。

+++ ChanEst_BCH_simplified/simplified_ChanEst_BCH/typedefs.h
@@ -46,18 +46,18 @@ typedef struct {
 } struct_creal;
 
 typedef struct {
-	double NDLRB;
+	int NDLRB;
 	char DuplexMode[20];
 	char CyclicPrefix[20];
-	double NCellID;
-	double NSubframe;
-	double TDDConfig;
-	double SSC;
-	double CellRefP;
+	int NCellID;
+	int NSubframe;
+	int TDDConfig;
+	int SSC;
+	int CellRefP;
 	char PHICHDuration[20];
 	char Ng[20];
-	double NFrame;
-	double CFI;
+	int NFrame;
+	int CFI;
 } struct_ENB;
 
 extern struct_creal crealAdd(struct_creal a, struct_creal b);
